life-os/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.36
pydantic-settings==2.2.1
python-dotenv==1.0.1
authlib==1.3.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
itsdangerous==2.2.0
google-api-python-client==2.125.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gradescopeapi==0.1.0
spotipy==2.24.0